Add rails default output
Since we're overriding rails default `self.scope` method, this commit
captures the output in a variable, and returns it at the end of the
method.
This would allow other gems or programs that rely on the default output
to perform any desired business logic or side effects, by overriding
themselves the `self.scope` method.