Fix errors access in rails 6
To make errors management compatible for rails 6.0, they're
now accessed in a way that will be compatible for both rails 6.0.x and
future versions. The `#each` API will have breaking changes starting
from rails 6.2 and since only a small subset of that API is used, the
`#messages` interface is used instead.